+++ source/newtonian/two_dimensional/source_terms/SimpleGravity.cpp
@@ -18,7 +18,7 @@ double /*dt*/)
 {
 	Vector2D pos(tess.GetCellCM(point) - (abs(_center)*Vector2D(cos(omega_*time),sin(omega_*time))));
 	double r = abs(pos);
-	double m = tess.GetVolume(point)*cells[point].Density;
+	double m = tess.GetVolume(point)*cells[static_cast<size_t>(point)].Density;
 	Vector2D acc;
 	if (r<softlength_)
 		acc=(-1)*pos*M_ / (r*r*r + Rmin_*Rmin_*r);
@@ -40,7 +40,7 @@ double /*dt*/)
 		else
 			dt_ = std::min(dt_, sqrt(tess.GetWidth(point) / abs(acc)));
 	}
-	return Conserved(0, m*acc, m*ScalarProd(acc, cells[point].Velocity));
+	return Conserved(0, m*acc, m*ScalarProd(acc, cells[static_cast<size_t>(point)].Velocity));
 }
 
 double SimpleGravity::GetTimeStep(void) const
